quiz_lab-2 Basic folder structure setup
